+++ docs/CHANGELOG.md
@@ -15,6 +15,7 @@ The entire project (i.e., the GraphQL API, REST API, and server) uses the same v
 ### Changed
 
 - Renamed `chatId` to `id` in `type UnstarredChat`.
+- `scalar Username`
 
 ## [0.17.0](https://github.com/neelkamath/omni-chat-backend/releases/tag/v0.17.0) - 2021-03-19